add ValidationTableViewProvider for GitHub Signup
