--- Example/Example/GitHubSigup/BindingExtensions.swift
@@ -1,69 +1,0 @@
-//
-//  BindingExtensions.swift
-//  RxExample
-//
-//  Created by Krunoslav Zaher on 12/6/15.
-//  Copyright © 2015 Krunoslav Zaher. All rights reserved.
-//
-
-import UIKit
-#if !RX_NO_MODULE
-import RxSwift
-import RxCocoa
-#endif
-
-extension ValidationResult: CustomStringConvertible {
-    var description: String {
-        switch self {
-        case let .ok(message):
-            return message
-        case .empty:
-            return ""
-        case .validating:
-            return "validating ..."
-        case let .failed(message):
-            return message
-        }
-    }
-}
-
-struct ValidationColors {
-    static let okColor = UIColor(red: 138.0 / 255.0, green: 221.0 / 255.0, blue: 109.0 / 255.0, alpha: 1.0)
-    static let errorColor = UIColor.red
-}
-
-extension ValidationResult {
-    var textColor: UIColor {
-        switch self {
-        case .ok:
-            return ValidationColors.okColor
-        case .empty:
-            return UIColor.black
-        case .validating:
-            return UIColor.black
-        case .failed:
-            return ValidationColors.errorColor
-        }
-    }
-}
-
-extension UniqueMessageTableViewProvider {
-    
-    var validationResult: Binder<ValidationResult> {
-        return Binder(self) { provider, result in
-            provider.backgroundView.backgroundColor = result.textColor
-            provider.messageLabel.text = result.description
-            provider.isHidden.accept(result.description.isEmpty)
-        }
-    }
-    
-}
-
-extension Reactive where Base: UILabel {
-    var validationResult: Binder<ValidationResult> {
-        return Binder(base) { label, result in
-            label.textColor = result.textColor
-            label.text = result.description
-        }
-    }
-}
